backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.7
python-jwt==4.0.0